+++ ethsigner/core/src/main/java/tech/pegasys/ethsigner/core/signing/TransactionSerialiser.java
@@ -12,7 +12,7 @@
  */
 package tech.pegasys.ethsigner.core.signing;
 
-import tech.pegasys.ethsigner.core.requesthandler.sendtransaction.Transaction;
+import tech.pegasys.ethsigner.core.requesthandler.sendtransaction.transaction.Transaction;
 
 import org.web3j.crypto.Sign.SignatureData;
 import org.web3j.crypto.TransactionEncoder;
